S115 cleanup and speedup, threading in S11

+++ helpers/helpers.sh
@@ -80,12 +80,13 @@ max_pids_protection() {
 }
 
 cleaner() {
-  print_output "[*] Ctrl+C detected!" "no_log"
+  print_output "[*] User interrupt detected!" "no_log"
   print_output "[*] Final cleanup started." "no_log"
-  # now we can unmount the stuff from emulator and delete temporary stuff
 
   # if S115 is found only once in main.log the module was started and we have to clean it up
-  if [[ -f "$LOG_DIR"/"$MAIN_LOG_FILE" ]]; then
+  # additionally we need to check some variable from a running emba instance
+  # otherwise the unmounter runs crazy in some corner cases
+  if [[ -f "$LOG_DIR"/"$MAIN_LOG_FILE" && "${#FILE_ARR[@]}" -gt 0 ]]; then
     if [[ $(grep -c S115 "$LOG_DIR"/"$MAIN_LOG_FILE") -eq 1 ]]; then
       print_output "[*] Terminating qemu processes - check it with ps" "no_log"
       killall -9 --quiet -r .*qemu.*sta.*
@@ -93,6 +94,7 @@ cleaner() {
       find "$FIRMWARE_PATH_CP" -xdev -iname "qemu*static" -exec rm {} \; 2>/dev/null
       print_output "[*] Umounting proc, sys and run" "no_log"
       mapfile -t CHECK_MOUNTS < <(mount | grep "$FIRMWARE_PATH_CP")
+      # now we can unmount the stuff from emulator and delete temporary stuff
       for MOUNT in "${CHECK_MOUNTS[@]}"; do
         print_output "[*] Unmounting $MOUNT" "no_log"
         MOUNT=$(echo "$MOUNT" | cut -d\  -f3)
